+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - The bloblang method `hash` and the `hash` processor now support `md5`.
 - Field `collector_url` added to the `jaeger` tracer.
 - The bloblang method `strip_html` now allows you to specify a list of allowed elements.
+- New bloblang method `parse_xml`.
 
 ## 3.37.0 - 2021-01-06